yolo-service/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
ultralytics==8.3.57
pydantic-settings==2.7.1
python-multipart==0.0.20
pillow==11.1.0
pytest==8.3.4
httpx==0.28.1
